# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.1",
    "title": "Astar token statistics API",
    "description": "Provides Astar networks statistic information."
  },
  "host": "api.astar.network",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/v1/{network}/dapps-staking/apr": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves dapp staking APR for a given network.",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden, shibuya, rocstar",
            "enum": [
              "astar",
              "shiden",
              "shibuya",
              "rocstar"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/{network}/dapps-staking/apy": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves dapp staking APY for a given network.",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden, shibuya, rocstar",
            "enum": [
              "astar",
              "shiden",
              "shibuya",
              "rocstar"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/{network}/dapps-staking/earned/{address}": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves earned staking rewards for dapps staking",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden, shibuya, rocstar",
            "enum": [
              "astar",
              "shiden",
              "shibuya",
              "rocstar"
            ]
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Wallet address. Supported address format: SS58"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/{network}/dapps-staking/dapps": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves list of dapps (full model) registered for dapps staking",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden, shibuya, rocstar, development",
            "enum": [
              "astar",
              "shiden",
              "shibuya",
              "rocstar"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/{network}/dapps-staking/dappssimple": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves list of dapps (basic info) registered for dapps staking",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden, shibuya, rocstar, development",
            "enum": [
              "astar",
              "shiden",
              "shibuya",
              "rocstar"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/{network}/dapps-staking/dapps/{address}": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves dapp with the given address",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden, shibuya, rocstar, development",
            "enum": [
              "astar",
              "shiden",
              "shibuya",
              "rocstar"
            ]
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The dapp address"
          },
          {
            "name": "forEdit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/{network}/dapps-staking/register": {
      "post": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Registers a new dapp",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden, shibuya, rocstar, development",
            "enum": [
              "astar",
              "shiden",
              "shibuya",
              "rocstar",
              "development"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/{network}/dapps-staking/stats/dapp/{contractAddress}/{period}": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves number of calls and unique users per era statistics.",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden",
            "enum": [
              "astar",
              "shiden"
            ]
          },
          {
            "name": "contractAddress",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contract address to get stats for"
          },
          {
            "name": "period",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Period to get stats for. Supported periods: 7 eras, 30 eras, 90 eras, all",
            "enum": [
              "7 eras",
              "30 eras",
              "90 eras",
              "all"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/{network}/dapps-staking/stats/user/{userAddress}/{period}": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves user transactions.",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden",
            "enum": [
              "astar",
              "shiden"
            ]
          },
          {
            "name": "userAddress",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User address to get stats for"
          },
          {
            "name": "period",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The period type. Supported values: 7 days 30 days, 90 days, 1 year",
            "enum": [
              "7 days",
              "30 days",
              "90 days",
              "1 year"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}